Update README.md
The project is about black money transactions to identify factors involved in money laundering, mapping the graphical flow of funds and developing a model to identify suspicious transactions. In the repository regular commits will be made to show the progress of project and that will help in maintaining our project.